Refactor chat into System Notification System
